feat: add mainChain Tsm Seed Info (#2)
feat:modify offerId
feat:add query method for listing、offer

+++ src/Forest.Indexer.Plugin/Processors/OfferRemovedLogEventProcessor.cs
@@ -5,6 +5,7 @@ using Forest.Indexer.Plugin.Entities;
 using Forest.Indexer.Plugin.Processors.Provider;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Nest;
 using Volo.Abp.ObjectMapping;
 
 namespace Forest.Indexer.Plugin.Processors;
@@ -22,10 +23,12 @@ public class OfferRemovedLogEventProcessor : OfferLogEventProcessorBase<OfferRem
         INFTOfferProvider offerProvider,
         ICollectionProvider collectionProvider,
         ICollectionChangeProvider collectionChangeProvider,
-        IOptionsSnapshot<ContractInfoOptions> contractInfoOptions) : base(logger, objectMapper,
+        IOptionsSnapshot<ContractInfoOptions> contractInfoOptions,
+        INFTOfferChangeProvider nftOfferChangeProvider) : base(logger, objectMapper,
         nftActivityIndexRepository, nftInfoIndexRepository, proxyAccountIndexRepository, infoProvider, offerProvider,collectionProvider,
         collectionChangeProvider,
-        contractInfoOptions)
+        contractInfoOptions,
+        nftOfferChangeProvider)
     {
         _nftOfferIndexRepository = nftOfferIndexRepository;
     }
@@ -37,29 +40,49 @@ public class OfferRemovedLogEventProcessor : OfferLogEventProcessorBase<OfferRem
 
     protected override async Task HandleEventAsync(OfferRemoved eventValue, LogEventContext context)
     {
-        var offerIndexId = IdGenerateHelper.GetId(context.ChainId, eventValue.Symbol, eventValue.OfferFrom.ToBase58(),
-            eventValue.OfferTo.ToBase58(), eventValue.ExpireTime.Seconds);
-        var offerIndex = await _nftOfferIndexRepository.GetFromBlockStateSetAsync(offerIndexId, context.ChainId);
-        if (offerIndex == null) return;
-        var nftInfoId = offerIndex.BizInfoId;
-        _objectMapper.Map(context, offerIndex);
-        await _nftOfferIndexRepository.DeleteAsync(offerIndex);
+        
+        var mustQuery = new List<Func<QueryContainerDescriptor<OfferInfoIndex>, QueryContainer>>();
+        mustQuery.Add(q => q.Term(i =>
+            i.Field(f => f.ChainId).Value(context.ChainId)));
+        mustQuery.Add(q => q.Term(i =>
+            i.Field(f => f.BizSymbol).Value(eventValue.Symbol)));
+        mustQuery.Add(q => q.Term(i =>
+            i.Field(f => f.OfferFrom).Value(eventValue.OfferFrom.ToBase58())));
+        mustQuery.Add(q => q.Term(i =>
+            i.Field(f => f.OfferTo).Value(eventValue.OfferTo.ToBase58())));
+        mustQuery.Add(q => q.Term(i =>
+            i.Field(f => f.ExpireTime).Value(eventValue.ExpireTime.ToDateTime())));  
 
-        var latestNFTOfferDic =
-            await _offerProvider.QueryLatestNFTOfferByNFTIdsAsync(new List<string> { offerIndex.BizInfoId },
+        QueryContainer ListingFilter(QueryContainerDescriptor<OfferInfoIndex> f) =>
+            f.Bool(b => b.Must(mustQuery));
+
+        var offerIndexList = await _nftOfferIndexRepository.GetListAsync(ListingFilter);
+        if (offerIndexList.Item1 == 0) return;
+        foreach (var index in offerIndexList.Item2)
+        {
+            var offerIndex = await _nftOfferIndexRepository.GetFromBlockStateSetAsync(index.Id, context.ChainId);
+            if (offerIndex == null) return;
+            var offerIndexId = offerIndex.Id;
+            var nftInfoId = offerIndex.BizInfoId;
+            _objectMapper.Map(context, offerIndex);
+            await _nftOfferIndexRepository.DeleteAsync(offerIndex);
+
+            var latestNFTOfferDic =
+                await _offerProvider.QueryLatestNFTOfferByNFTIdsAsync(new List<string> { offerIndex.BizInfoId },
+                    offerIndexId);
+
+            var latestNFTOffer = latestNFTOfferDic != null && latestNFTOfferDic.ContainsKey(offerIndex.BizInfoId)
+                ? latestNFTOfferDic[offerIndex.BizInfoId]
+                : new OfferInfoIndex()
+                {
+                    BizInfoId = nftInfoId
+                };
+
+            await _infoProvider.UpdateOfferCommonAsync(context.ChainId, eventValue.Symbol, context,
+                latestNFTOffer,
                 offerIndexId);
-
-        var latestNFTOffer = latestNFTOfferDic != null && latestNFTOfferDic.ContainsKey(offerIndex.BizInfoId)
-            ? latestNFTOfferDic[offerIndex.BizInfoId]
-            : new OfferInfoIndex()
-            {
-                BizInfoId = nftInfoId
-            };
-
-        await _infoProvider.UpdateOfferCommonAsync(context.ChainId, eventValue.Symbol, context,
-            latestNFTOffer,
-            offerIndexId);
+        }
         await _collectionChangeProvider.SaveCollectionPriceChangeIndexAsync(context, eventValue.Symbol);
-
+        await _nftOfferChangeProvider.SaveNFTOfferChangeIndexAsync(context, eventValue.Symbol, EventType.Remove);
     }
 }
